[refactor:boundary] cn-protocol 共有 wire crate を抽出 (WP-H3 PR1)
デスクトップ ↔ コミュニティノード間の通信の形を新 crate
kukuri-cn-protocol へ抽出(依存: serde / serde_json / url / kukuri-core のみ):
- cn-core から移動: models.rs(wire 型一式)/ normalize.rs(URL・pubkey
  正規化、封筒検証)/ auth.rs の client 部分(build_auth_envelope_json +
  AUTH_ENVELOPE_KIND)
- desktop-runtime から移動: manifest の client 用 slim 型 4 struct
  (取得ロジックと round-trip 契約テストは残置)
- cn-core は re-export のシムで公開面を維持(サーバ側 import 無変更)
- desktop-runtime の依存を cn-core → cn-protocol へ差し替え。
  依存ツリーから axum / redis / jsonwebtoken が消滅(sqlx はローカル
  sqlite 用に store 経由で残存)

serde 表現は型の移動のみで不変(wire バイト互換)。
