+++ crates/cn-core/Cargo.toml
@@ -31,6 +31,7 @@ tracing.workspace = true
 url.workspace = true
 uuid.workspace = true
 kukuri-core = { path = "../core" }
+kukuri-cn-protocol = { path = "../cn-protocol" }
 kukuri-cn-safety = { path = "../cn-safety" }
 kukuri-cn-safety-arachnid = { path = "../cn-safety-arachnid", optional = true }
 kukuri-cn-safety-runtime = { path = "../cn-safety-runtime" }
