+++ crates/cn-protocol/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "kukuri-cn-protocol"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+url.workspace = true
+kukuri-core = { path = "../core" }
+
+[lints]
+workspace = true
